requirements.txt
asgiref>=3.10.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
django-cors-headers==4.9.0
gunicorn==23.0.0
whitenoise==6.6.0
python-socketio==5.11.1
eventlet==0.36.1
websocket-client==1.7.0
django-ratelimit==4.1.0
uvicorn==0.27.0
uvicorn[standard]==0.27.0
psutil==6.1.0
requests==2.31.0